Add notebooks for the separate and combined UIs
Demonstrate the individual UIs for gti.apps.descr and gti.apps.linked
in separate notebooks. Note that the Tabulator headers are visible and
that the UIs function correctly.

Demonstrate that, when these 2 UIs are put in separate tabs with a
GoldenTemplate layout and the Tabulators are linked, the headers
disappear.

Note that the watched function linking the Tabulators (selecting a
row in the Description UI tab updates the two tables in the Linked UI
tab) operates correctly.